Added Ontology Base component
